Update README.md
This is a dashboard showing the analysis of the covid 19. Gives the total number of confirmed cases, the recovered cases as well as the death rate and total number of deaths. I got my data from the web.I learnt this during the #30DaysOfLearning. I hope i have done it well for a first time and will be doing more.